requirements.txt
numpy
pandas
matplotlib
pyyaml
conda